Bugfix?
Deleting the modules in the same order as they where created may be a problem. A controller is loaded after a simulation core. In a multi-threaded environment, the controller may still access the simulation core while that module is already unloaded. As a workaround, unload the modules in reverse order.