Ensure that the main window is closed first on macOS
Otherwise, floating dock widgets could get a close event before
the main window, which means that the widget will not be deleted
in MainWindow::closeFile (because it unregistered itself in
MainWindow::closedObject) and thus only be deleted after the main
window when the modules have already been unloded.

At the moment, when the program is quit by closing the main window
directly, this leads to two close events for the main window, but
that doesn't hurt.