Workaround for Linux
If the QGtk3Dialog is used, gtk_dialog_run is called instead of a new
QEventLoop. This seems to handle the timer with interval 0 incorrectly,
i.e. the dialog never appears if the simulation is running (strangely,
the SimObjectWidget continues to be rendered, but the application does
not respond anymore).