Register model matrices according to their type
When rendering a camera image, only appearances' matrices need
to be updated. There is still a lot more to save in the updates
when drawing a SimObjectRenderer.

This also fixes a bug in the SimObjectRenderer where previously the
originPose would be based on a previous value of the object's pose.

For OpenGL, all of this could be much simpler (just calculating
the matrix in the draw call), but that can't be done with Vulkan
or Metal since the command buffers will be prerecorded there.